update codes, upgrade BLINKER PRO codes
